refactor: controller and service transaction
